requirements.txt
Django
PyYAML
coverage
colour-runner
lupa
psycopg2
rules
wait
parameterized
django-allauth
django-autocomplete-light
django-debug-toolbar
django-sortedm2m
django-nested-admin
django-registration
django-environ
# django-subdomains doesn't work with Django 1.10+ so we're using a commit from a pull request
-e git+https://github.com/huwshimi/django-subdomains.git@63405c239148308bf214f950c0cff50d7f29b23d#egg=django-subdomains
